requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
pkg_resources==0.0.0
protobuf==4.21.12
